Trial and error on Vec
The issue was to be able to use map on matrixes obtaining an Iterator from it. This caused its to be borrowed  in the scope etc etc and moved after...I think that the proper solution lies in using lifetimes and & to matrixes inside the RiderParameters as long as it does not need to change it but I'm still not able to do that.