Examples for a nasty bug about successive indels
For individuals with both dels mutated a window gets repeated because we
skip at the first base following the first del when we have a window starting with the first del but the second del eats away the base that here we would have considered.
Check window starts effective to avoid this? Or...?